config: ast2600: Add config for SPL MMC boot with FIT signature verification
This adds a configuration for FIT signature verification for the OpenBMC SPL
configuration, for both the SPL and u-boot. It uses the common MMC SPL
loader code present in u-boot.

This enables rsa and SHA512 algorithms for verifying the signature of the
u-boot FIT when loading it. Hashing is performed by the AST2600's HACE.

The SPL will load u-boot proper from sector 0x80 (128), which is 64KB
into the device.

The u-boot load size of 1MB, minus 64KB for the SPL, is configured through CONFIG_SYS_MONITOR_LEN. FIT
can describe it's length, so this is only used when loading a raw image.

Ensure the SPL fits in the 64KB payload size with FIT signature
verification in the SPL.

 - Build with thumb

 - Disable the arch memcpy/memset changes to save 668 bytes. The tiny
   memset saves 52 bytes. (Anyone who wants to test the runtime impact
   of these changes should do so!)

 - FIT_IMAGE_TINY is selected to save approx 3KB from the image size.

Set the size of the early (pre-DRAM) SRAM heap in u-boot proper to 8KB.
The DRAM driver would perform an allocation that exceeded the default
limit, due to the probing of the HACE driver increasing memory
presssure. As SRAM is unused when u-boot proper is running, it can all
be allocated.

(In theory the entire 88KB of SRAM is free, but testing showed more than
60KB would break booting. Finding out why is TODO).

The SPL early heap is fixed to 0x800, the default before this change.
Testing shows that:

  SPL malloc() before relocation used 0x794 bytes (1 KB)

So 2KB should be enough.

Signed-off-by: Joel Stanley <joel@jms.id.au>