crypto: Restore old driver for Aspeed HACE
The HACE supports MD5, SHA1 and SHA2 family hash functions. This driver
uses it in a polling mode to perform hash calculations over buffers
placed in DRAM.

It also includes a workaround[1] that performs a reset when error occurs.
This is incompatible with simultaneous operation of the ACRY
hardware, as the HACE reset line is shared with the ACRY hardware. It
should only be used as a workaround for when the ASPEED_HACE driver,
which uses accumulative mode, cannot be used.

 [1] https://lore.kernel.org/r/20210715045258.13912-1-joel@jms.id.au

Currently simulator for support for accumulative mode is not in place,
so this driver exists for supporting machines that must boot in a
simulator.

Signed-off-by: Joel Stanley <joel@jms.id.au>