ARM: dts: aspeed: Add a Romulus board (AST2500)
Romulus was the OpenPOWER reference platform for the IBM Power9,
containing an ASPEED AST2500.

It was designed by IBM and built by SuperMicro. This reference platform
was sold as the Raptor Computing Systems Tallos II.

The machine used one 32MB SPI NOR for the BMC, a second SPI NOR flash
for the host firmware, and the Broadcom BCM5719 attached via NC-SI.

Signed-off-by: Joel Stanley <joel@jms.id.au>
Reviewed-by: Cédric Le Goater <clg@kaod.org>
Link: https://lore.kernel.org/r/20220804035504.92900-1-joel@jms.id.au